% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g1979a1}
\alias{g1979a1}
\title{Glushko (1979), Word & derived nonword sets for regular and exception items.}
\format{A data.frame with 172 rows and 6 variables:
\describe{
    \item{set}{Item set.}

    \item{ortho}{Orthographic representation.}

    \item{phono}{Phonological representation. Not included in original
    source.}

    \item{type}{Word or nonword.}

    \item{rt_ms}{Correct pronunciation latencies.}

    \item{regularity}{Spelling regularity bin: R (regular)
    or E (exception).}
}}
\source{
Glushko, R. J. (1979). The organization and activation of orthographic
    knowledge in reading aloud. Journal of Experimental Psychology: Human
    Perception and Performance, 5(4), 674-691. doi:10.1037/0096-1523.5.4.674
}
\usage{
g1979a1
}
\description{
This data set gathered from table A1 includes 172 monosyllabic English
    words and lexically derived nonwords (43 groups). Within each group, there are two
    pairs of items that are either regular or exception words or nonwords.  Each regular
    and exception word has a homophonically similar nonword counterpart that is matched
    for onset phonome but occasionally contains irregular spelling-to-sound correspondences.
}
\details{
No details yet.
}
\seealso{
\code{\link{g1979a2}} \code{\link{g1979a3}}
}
\keyword{datasets}

