% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g1979a2}
\alias{g1979a2}
\title{Glushko (1979), Lexically derived nonword set.}
\format{A data.frame with 52 rows and 5 variables:
\describe{
    \item{set}{Item set.}

    \item{ortho}{Orthographic representation.}

    \item{phono}{Phonological representation. Not included in original
    source.}

    \item{rt_ms}{Correct pronunciation latencies.}

    \item{regularity}{Spelling regularity bin: R (regular) or E
    (exception).}
}}
\source{
Glushko, R. J. (1979). The organization and activation of orthographic
    knowledge in reading aloud. Journal of Experimental Psychology: Human
    Perception and Performance, 5(4), 674-691. doi:10.1037/0096-1523.5.4.674
}
\usage{
g1979a2
}
\description{
This data set gathered from table A2 includes a list of 52 lexically
    derived monosyllabic nonwords (26 pairs). Source words are either Regular or
    exception words.
}
\details{
See source paper for details.
}
\seealso{
\code{\link{g1979a1}} \code{\link{g1979a3}}
}
\keyword{datasets}
