% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g1979a3}
\alias{g1979a3}
\title{Glushko (1979), Pronunciation Latency by Consitency and Regularity
    word set.}
\format{A data.frame with 164 rows and 5 variables:
\describe{
    \item{set}{Item set.}

    \item{ortho}{Orthographic representation.}

    \item{rt_ms}{correct pronunciation latencies}

    \item{consistency}{}

    \item{regularity}{Spelling regularity bin: R (regular) or E
    (exception).}
}}
\source{
Glushko, R. J. (1979). The organization and activation of orthographic
    knowledge in reading aloud. Journal of Experimental Psychology: Human
    Perception and Performance, 5(4), 674-691. doi:10.1037/0096-1523.5.4.674
}
\usage{
g1979a3
}
\description{
This data set gathered from table A3 includes 164 monosyllabic
    English words (41 word groups).  Within each group, there are two consistent
    words that contrast in regularity (E vs. R) and two inconsistent words that
    contrast in regularity (E vs. R).  Inconsistent words within each grouping
    contrast in onset phoneme while consistent words contrast in onset phoneme as
    well as end phoneme.
}
\details{
See source paper for details.
}
\seealso{
\code{\link{g1979a2}} \code{\link{g1979a1}}
}
\keyword{datasets}
