% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mb1987a}
\alias{mb1987a}
\title{McCann & Besner (1987), Word and derived nonword/pseudohomophone set.}
\format{A data.frame with 240 rows and 8 variables:
\describe{
    \item{set}{Item set. Lines up with 'set' in \code{\link{mb1987b}}}

    \item{ortho}{Orthographic representation.}

    \item{phono}{Phonological representation. Not included in original
    source.}

    \item{type}{Item type: "word", "nonword" or "pseudohom"}

    \item{OPC_RT}{}

    \item{DPC_RT}{}

    \item{OPC_ERR}{}

    \item{DPC_ERR}{}
}}
\source{
McCann, R. S., & Besner, D. (1987). Reading pseudohomophones:
    Implications for models of pronunciation assembly and the locus
    of word-frequency effects in naming. Journal of Experimental Psychology:
    Human Perception and Performance, 13(1), 14-24.
    doi:10.1037/0096-1523.13.1.14
}
\usage{
mb1987a
}
\description{
This data set gathered from appendix A includes 240 Word and
    derived nonword/pseudohomophones (80 triads). Includes average reaction time data.
}
\details{
No details yet.
}
\seealso{
\code{\link{mb1987b}}
}
\keyword{datasets}
