% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mb1987b}
\alias{mb1987b}
\title{McCann & Besner (1987), Reaction Time and Error Rate based on
    Nonwords and Nonword PSeudohomophones.}
\format{A data.frame with 160 rows and 6 variables:
\describe{
    \item{set}{Item set. Lines up with 'set' in \code{\link{mb1987a}} }

    \item{ortho}{Orthographic representation.}

    \item{phono}{Phonological representation. Not included in original
    source.}

    \item{RT}{Reaction time.}

    \item{error}{Percentage of error response.}

    \item{src_type}{Type of source item.}
}}
\source{
McCann, R. S., & Besner, D. (1987). Reading pseudohomophones:
    Implications for models of pronunciation assembly and the locus
    of word-frequency effects in naming. Journal of Experimental Psychology:
    Human Perception and Performance, 13(1), 14-24.
    doi:10.1037/0096-1523.13.1.14
}
\usage{
mb1987b
}
\description{
This data set gathered from appendix B includes 160 nonword
    pseudohomophones (80 nonword pairs). Includes average reaction time
    data and error rate.
}
\details{
No details yet.
}
\seealso{
\code{\link{mb1987a}}
}
\keyword{datasets}

