% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rgm2002c}
\alias{rgm2002c}
\title{Rodd, Gaskell & Marslen-Wilson (2002), Senses by Ambiguity word set.}
\format{A data.frame with 92 rows and 3 variables:
\describe{
    \item{ortho}{Orthographic representation.}

    \item{senses}{The number of distinct semantic 'senses' for each word.}

    \item{ambiguous}{Whether or not the word has more than one
    meaning: Y (yes) or N (no)}
}}
\source{
Rodd, J., Gaskell, G., & Marslen-Wilson, W. (2002). Making Sense
    of Semantic Ambiguity: Semantic Competition in Lexical Access. Journal
    of Memory and Language, 46(2), 245-266. doi:10.1006/jmla.2001.2810
}
\usage{
rgm2002c
}
\description{
This data set gathered from appendix A includes 92 English
    words.  Source words are either ambiguous or unambiguous and vary in number
    of senses.
}
\details{
No details yet.
}
\seealso{
\code{\link{rgm2002a}} \code{\link{rgm2002b}}
}
\keyword{datasets}
