% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sps1995a}
\alias{sps1995a}
\title{Strain, Patterson & Seidenberg (1995), Frequency by Regularity by Imageability.}
\format{A data.frame with 96 rows and 4 variables:
\describe{
    \item{ortho}{The orthographic word.}

    \item{frequency}{Word frequency bin: H (high) or L (low).}

    \item{regularity}{Spelling regularity bin: R (regular) or E
    (exception).}

    \item{imageability}{Imageability bin: H (high) or L (low).}
}}
\source{
Strain, E., Patterson, K., & Seidenberg, M. S. (1995). Semantic effects
    in single-word naming. Journal of Experimental Psychology: Learning, Memory,
    and Cognition, 21(5), 1140-1154. doi:10.1037/0278-7393.21.5.1140
}
\usage{
sps1995a
}
\description{
This data set gathered from appendix A includes 96 English words of
    differing syllabic structure.  Source words are either regular or exception
    words that have either high or low frequency and imageability.
}
\details{
See source paper for details.
}
\seealso{
\code{\link{sps1995b}}
}
\keyword{datasets}
