% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sps1995b}
\alias{sps1995b}
\title{Strain, Patterson & Seidenberg (1995), Regularity by Imageability in low Frequency words.}
\format{A data.frame with 64 rows and 4 variables:
\describe{
    \item{ortho}{The orthographic word.}

    \item{frequency}{word frequency bin: L (low).}

    \item{regularity}{Spelling regularity bin: R (regular) or E
    (exception).}

    \item{imageability}{Imageability bin: H (high) or L (low).}
}}
\source{
Strain, E., Patterson, K., & Seidenberg, M. S. (1995). Semantic effects
    in single-word naming. Journal of Experimental Psychology: Learning, Memory,
    and Cognition, 21(5), 1140-1154. doi:10.1037/0278-7393.21.5.1140
}
\usage{
sps1995b
}
\description{
This data set gathered from appendix B includes 64 English words of differing
    syllabic structure. Source words are either regular or exception words that are all low
    frequency but have either high or low imageability.
}
\details{
No details yet.
}
\seealso{
\code{\link{sps1995a}}
}
\keyword{datasets}

