% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sps2002b}
\alias{sps2002b}
\title{Strain, Patterson & Seidenberg (2002), Regularity word set.}
\format{A data.frame with 120 rows and 3 variables:
\describe{
    \item{word}{The orthographic word.}

    \item{set}{Item set. Maps word pairs, contrasting in regularity, but
    approximately matched for onset phoneme.}

    \item{regularity}{Spelling regularity bin: R (regular) or E
    (exception).}
}}
\source{
Strain, E., Patterson, K., & Seidenberg, M. S. (2002). Theories of word
    naming interact with spelling--sound consistency. Journal of Experimental
    Psychology: Learning, Memory, and Cognition, 28(1), 207-214.
    doi:10.1037/0278-7393.28.1.207
}
\usage{
sps2002b
}
\description{
This data set gathered from appendix B includes 120 monosyllabic
    English words (60 word pairs). Within each pair, words contrast in regularity
    (R vs. E) but are matched for onset phoneme.
}
\details{
No details yet.
}
\keyword{datasets}

