% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tgb1990}
\alias{tgb1990}
\title{Treiman et al (1990), Nonword set.}
\format{A data.frame with 48 rows and 3 variables:
\describe{
    \item{nonwords}{Orthographic representation of nonwords.}

    \item{phon_rep}{Phonological representation with adjusted phonological
     symbols.}

    \item{frequency}{Word frequency bin: H (high) or L (low).}
}}
\source{
Treiman, R., Goswami, U., & Bruck, M. (1990). Not all nonwords are
    alike: Implications for reading development and theory. Memory & Cognition,
    18(6), 559-567. doi:10.3758/BF03197098
}
\usage{
tgb1990
}
\description{
This data set gathered from the source appendix includes
    48 nonwords with either high or low frequency.
}
\details{
No details yet.
}
\keyword{datasets}
