% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tm1987a1}
\alias{tm1987a1}
\title{Taraban & McClelland (1987), Frequency by Regularity word set.}
\format{A data.frame with 96 rows and 8 variables:

\describe{

    \item{ortho}{The orthographic word.}

    \item{set}{Item set. Maps word pairs, contrasting in regularity, but
    approximately matched for onset phoneme, word frequency, and
    length.}

    \item{wordfreq}{Word frequency.}

    \item{samepron}{Number of other words with same orthographic word
    body having the same pronunciation.}

    \item{MFA}{Number of other words where the same orthographic word
    body has the Most Frequent Alternative pronunciation.}

    \item{LFA}{Number of other words where the same orthographic word
    body has a Less Frequent Alternative pronunciation.}

    \item{frequency}{Word frequency bin: H (high) or L (low).}

    \item{regularity}{Spelling regularity bin: R (regular) or E
    (exception).}

}}
\source{
Taraban, R., & McClelland, J. L. (1987). Conspiracy effects
    in word pronunciation. Journal of Memory and Language, 26(6), 608-631.
    doi:Doi 10.1016/0749-596x(87)90105-7
}
\usage{
tm1987a1
}
\description{
This data set gathered from appendix A and includes
    96 monosyllabic English words (48 word pairs). Within each pair,
    words contrast in regularity (R vs. E), but are approximately
    matched for onset phoneme, word frequency, and length.
}
\details{
No details yet.
}
\seealso{
\code{\link{tm1987a2}}
}
\keyword{datasets}
