% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tm1987a2}
\alias{tm1987a2}
\title{Taraban & McClelland (1987), Frequency by Consistency word set.}
\format{A data.frame with 96 rows and 10 variables:
\describe{
    \item{ortho}{The orthographic word.}

    \item{set}{Item set. Maps word pairs, contrasting in consistency, but
    approximately matched for onset phoneme, word frequency, length
    and regularity.}

    \item{altpron}{Alternate pronunciations of inconsistent words.}

    \item{wordfreq}{Word Frequency.}

    \item{samepron}{Number of words with same orthographic word
    body having the same pronunciation.}

    \item{MFA}{Number of words where the same orthographic word
    body has the most frequent alternative pronunciation.}

    \item{LFA}{Number of words where the same orthographic word
    body has a less frequent alternative pronunciation.}

    \item{frequency}{Word frequency bin: H (high) or L (low).}

    \item{regularity}{Spelling regularity bin: R (regular)}

    \item{consistency}{}
}}
\source{
Taraban, R., & McClelland, J. L. (1987). Conspiracy effects
    in word pronunciation. Journal of Memory and Language, 26(6), 608-631.
    doi:Doi 10.1016/0749-596x(87)90105-7
}
\usage{
tm1987a2
}
\description{
This data set gathered from appendix A includes
    96 monosyllabic English words (48 word pairs). Each pair includes
    alternate pronunciations for provided inconsistent words.  Within each
    pair words contrast in consistency (I vs. C), but are approximately
    matched for onset phoneme, word frequency, length and regularity.
}
\details{
See source paper for details.
}
\seealso{
\code{\link{tm1987a1}}
}
\keyword{datasets}
